moved a function, added a permissions check

+++ src/periwinkle/store/group.go
@@ -156,32 +156,6 @@ func (o *Group) Save(db *gorm.DB) {
 	}
 }
 
-func IsSubscribed(db *gorm.DB, userid string, group Group) bool {
-	subscriptions := GetSubscriptionsGroupById(db, group.Id)
-	address_ids := make([]int64, len(subscriptions))
-	for i, subscription := range subscriptions {
-		address_ids[i] = subscription.AddressId
-	}
-	var addresses []UserAddress
-	if len(address_ids) > 0 {
-		if result := db.Where("id IN (?)", address_ids).Find(&addresses); result.Error != nil {
-			if !result.RecordNotFound() {
-				panic("cant find any subscriptions corresponding user address")
-			}
-		}
-	} else {
-		// no subscriptions so user cannot possibly be subscribed
-		return false
-	}
-	for _, address := range addresses {
-		if address.UserId == userid {
-			return true
-		}
-	}
-	// could not find user in subscribed user addresses, therefore, he/she isn't subscribed
-	return false
-}
-
 func (o *Group) Subentity(name string, req he.Request) he.Entity {
 	panic("TODO: API: (*Group).Subentity()")
 }
@@ -234,6 +208,10 @@ func (o *Group) Methods() map[string]func(he.Request) he.Response {
 		},
 		"DELETE": func(req he.Request) he.Response {
 			db := req.Things["db"].(*gorm.DB)
+			sess := req.Things["session"].(*Session)
+			if !IsAdmin(db, sess.UserId, *o) {
+				return he.StatusForbidden(heutil.NetString("Unauthorized user"))
+			}
 			db.Delete(o)
 			return he.StatusNoContent()
 		},
